Tons of task serialization support and some improvements, needs testing. Hopefully it's great!
Former-commit-id: e8bcff7c010c301ce794473f659c91ae20d8a571 [formerly 32d033f2668f87de79a0919aaa7903737d7709e0]
Former-commit-id: 68ca5fafe621fbff07e7629a69387049fa9313c7

+++ Source/Game/NetworkLayer/lobbyServer.h
@@ -149,11 +149,11 @@ public:
 		return ourUrl;
 	}
 
-	uint32_t port() const {
+	uint16_t port() const {
 		return ourPort;
 	}
 
-	void setEndpoint(const std::string &a_url, uint32_t a_port) {
+	void setEndpoint(const std::string &a_url, uint16_t a_port) {
 		activeState = AVAILABLE;
 		ourUrl = a_url;
 		ourPort = a_port;
@@ -182,7 +182,7 @@ private:
 	LobbyServer& ourServer;
 
 	std::string ourUrl;
-	uint32_t ourPort;
+	uint16_t ourPort;
 
 	std::shared_ptr<MatchSeeker> leftPlayer;
 	std::shared_ptr<MatchSeeker> rightPlayer;
